chore(test.yml): add maven settings to look for dependencies in the Liquibase repositories
feat(test.yml): add Liquibase repositories and servers configuration to enable access to Liquibase dependencies

The changes in this commit add the necessary configuration to the test workflow to look for dependencies in the Liquibase repositories. This is achieved by adding a maven-settings-xml-action step that configures the repositories and servers for Liquibase. The repositories are configured with the URLs for the Liquibase and Liquibase Pro repositories, enabling access to their releases and snapshots. The servers are configured with the necessary credentials to authenticate with the Liquibase and Liquibase Pro repositories. This allows the workflow to install the liquibase-core 0-SNAPSHOT dependency.